requirements.txt
tensorflow==2.14.0
absl-py
numpy
scipy
tensorflow_addons
